refactors library docgen code into docgen_lib
